HANDLERS: More work in the structure of handlers.
- New file 'base.py' to include all the common methods.
- Moved some handlers to runder.py (work in progress).
- Deleted handlers.py.
