+++ dbod/api/rundeck.py
@@ -12,8 +12,6 @@
 REST API Server for the DB On Demand System
 """
 
-from dbod.api.dbops import *
-from dbod.config import config
 import tornado.web
 import tornado.log
 import base64
@@ -22,65 +20,7 @@ import logging
 import json
 import requests
 
-
-# HTTP API status codes
-OK = 200
-CREATED = 201 # Request fulfilled resulting in creation of new resource
-NO_CONTENT = 204 # Succesfull delete
-NOT_FOUND = 404
-UNAUTHORIZED = 401
-
-# Basic HTTP Authentication decorator
-def http_basic_auth(fun):
-    """Decorator for extracting HTTP basic authentication user/password pairs
-    from the request headers and matching them to configurated credentials.
-    It will generate an HTTP UNAUTHORIZED (Error code 401) if the request is 
-    not using HTTP basic authentication.
-
-    Example:
-        @http_basic_auth
-        def get(self, user, pwd)
-    """
-    @functools.wraps(fun)
-    def wrapper(*args, **kwargs):
-        """ Decorator wrapper """
-        self = args[0]
-        try:
-            # Try
-            auth = self.request.headers.get('Authorization')
-            scheme, _, token = auth.partition(' ')
-            if scheme.lower() == 'basic':
-                # Decode user and password
-                user, _, pwd = base64.decodestring(token).partition(':')
-                if user == config.get('api','user') and pwd == config.get('api','pass'):
-                    return fun(*args, **kwargs)
-                else:
-                    # Raise UNAUTHORIZED HTTP Error (401) 
-                    logging.error("Unauthorized access from: %s",
-                        self.request.headers)
-                    raise tornado.web.HTTPError(UNAUTHORIZED)
-
-            else:
-                # We only support basic authentication
-                logging.error("Authentication scheme not recognized")
-                return "Authentication scheme not recognized"
-        except AttributeError:
-            # Raise UNAUTHORIZED HTTP Error (401) if the request is not
-            # using autentication (auth will be None and partition() will fail
-            logging.error("Unauthorized access from: %s",
-                self.request.headers)
-            raise tornado.web.HTTPError(UNAUTHORIZED)
-
-    return wrapper
-
-class DocHandler(tornado.web.RequestHandler):
-    """Generates the API endpoint documentation"""
-    def get(self):
-        logging.info("Generating API endpoints doc")
-        response = """Please use :
-            <p>http://hostname:port/api/v1/entity/NAME</p>
-            <p>http://hostname:port/api/v1/host/HOSTNAME</p>"""
-        self.write(response)
+from dbod.config import config
 
 class RundeckResources(tornado.web.RequestHandler):
     def get(self):
